fix(desktop): keep local draft recovery actionable
